+++ Sources/PippinAdapters/DefaultEnvironment.swift
@@ -14,13 +14,13 @@ public extension Environment {
     /// Currently provides default instances of these adapters:
     ///  - `ActivityIndicator`: `JGProgressHUDAdapter`
     ///  - `Alerter`: `SwiftMessagesAdapter`
+    ///  - `BugReporter`: `PinpointKitAdapter`
+    ///  - `CrashReporter`: `KSCrashAdapter`
     ///  - `Logger`: `XCGLoggerAdapter`
     /// Any component may still be switched out afterwards as desired.
     ///
-    /// TODO: add a `CountlyAdapter` default for `CrashReporter`.
-    ///
     /// - Parameters:
-    ///     - bugReportRecipients: The email addresses to which bug reports should be sent. Optional; if provided, an instance of `PinpointKitAdapter` (`BugReporter`) is constructed.
+    ///     - bugReportRecipients: The email addresses to which bug and crash reports should be sent. Optional; if provided, instances of both `PinpointKitAdapter` (`BugReporter`) and `KSCrashAdapter` (`CrashReporter`) are constructed.
     ///     - touchVizRootVC: The root `UIViewController` that will be used to construct the app hierarchy to enable touch visualization. Optional; if provided, an instance of `COSTouchVisualizerAdapter` (`TouchVisualization`) is constructed.
     ///
     /// - Note: `CrashlyticsAdapter` may not be included due to it's static dependency and limitations of CocoaPods.
@@ -37,6 +37,7 @@ public extension Environment {
 
         if let bugReportRecipients = bugReportRecipients {
             environment.bugReporter = PinpointKitAdapter(recipients: bugReportRecipients)
+            environment.crashReporter = KSCrashAdapter(recipients: bugReportRecipients)
         }
 
         environment.connectEnvironment()