+++ Sources/PippinAdapters/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ## [Unreleased]
 
+### Added
+
+- Added an adapter for KSCrash as an alternative for Crashlytics.
+
 ## [6.0.0] 2020-10-26
 
 ### Changed